[CDN-1175] Pass correct args to delete_ssl_certificate flow.
When a domain from poppy service is deleted, update_service flow
will be invoked. This flow in turn calls delete_ssl_certificate
flow but does not pass required arguments flavor_id and
providers_list. This causes delete_ssl_certificate flow to fail.
Fixing this to supply the correct arguments to the flow.