[CDN-1210] Advanced cname generation.
When creating a new service, return an intermediate domain
name to the user at the time of service creation itself
instead of waiting for the cps certs availability.
The user can cname his domain to this returned value.

After returning the intermediate domain to user,
depending on the certs availability:

CASE 1:
	If the cps certs are available, point the
	intermediate domain to the available cps cert
CASE 2:
	If the cps certs are NOT available, point
	the intermediate domain to a temporary domain
	(read from config). Later, when any cps certs
	becomes available, update/replace the temporary
	domain with actual cps cert.