[CDN-1251] Modify retry_list API to support cert_type.
Currently the GET and PUT APIs of `retry_list` does not support the
parameter `cert_type` from the request payload.

This is an issue which can be observed by manually the calling these
APIs and supplying the `cert_type` parameter in the request payload
for PUT. The same issue is observed for GET request as well when
the backend programatically had `cert_type` in the retry queue but
GET will fail to fetch these details.

Modifying both GET and PUT requests to enable the support for
`cert_type` parameter.